Markdown: Allow markdown parser to be imported on its own
Fixes #935

This is particularly useful to homogenize server side markdown parsing
with client presentation.

Change-type: minor
