+++ src/extra/Markdown.tsx
@@ -1,57 +1,10 @@
-import marked from 'marked';
 import * as React from 'react';
-import sanitizeHtml from 'sanitize-html';
 import styled from 'styled-components';
 import Txt, { TxtProps } from '../components/Txt';
+import { parseMarkdown } from './utils';
 
 type MarkdownProps = TxtProps & {
 	children: string;
-};
-
-const renderer = new marked.Renderer();
-renderer.link = function(_href, _title, _text) {
-	const link = marked.Renderer.prototype.link.apply(this, arguments);
-	return link.replace('<a', '<a target="_blank" rel="noopener noreferrer"');
-};
-
-const markedOptions = {
-	// Enable github flavored markdown
-	gfm: true,
-	breaks: true,
-	headerIds: false,
-	// Input text is sanitized using `sanitize-html` prior to being transformed by
-	// `marked`
-	sanitize: false,
-	renderer,
-};
-
-const sanitizerOptions = {
-	allowedTags: sanitizeHtml.defaults.allowedTags.concat([
-		'del',
-		'h1',
-		'h2',
-		'img',
-		'input',
-	]),
-	allowedAttributes: {
-		a: ['href', 'name', 'target', 'rel'],
-		img: ['src'],
-		input: [
-			{
-				name: 'type',
-				multiple: false,
-				values: ['checkbox'],
-			},
-			'disabled',
-			'checked',
-		],
-	},
-};
-
-const render = (text: string = '') => {
-	const html = marked(text, markedOptions);
-	const clean = sanitizeHtml(html, sanitizerOptions as any);
-	return clean;
 };
 
 /*
@@ -762,7 +715,7 @@ export const Markdown = ({ children, ...props }: MarkdownProps) => {
 	return (
 		<GitHubMarkdown
 			{...props}
-			dangerouslySetInnerHTML={{ __html: render(children) }}
+			dangerouslySetInnerHTML={{ __html: parseMarkdown(children) }}
 		/>
 	);
 };
